requirements.txt
selenium==4.27.0
webdriver-manager==4.0.2
python-dotenv==1.0.1
requests==2.32.0
beautifulsoup4==4.12.3
